rename and change semantics of VALIDATED and NOCHAIN flags
  * rename the SCM_FLAG_VALIDATED flag to SCM_FLAG_VALID

  * rename the SCM_FLAG_NOCHAIN flag to SCM_FLAG_NOTVALID

  * change the semantics of SCM_FLAG_VALID to mean "the object is
    currently valid (if SCM_FLAG_NOTVALID is not set)" from the
    previous "the flag was valid at some point in the past and is
    still valid if SCM_FLAG_NOCHAIN is not set"

  * change the semantics of SCM_FLAG_NOTVALID to mean "the object is
    currently not valid" (trumps SCM_FLAG_VALID, if that is set) from
    the previous "there is no chain of valid or invalid certificates
    to a trust anchor"

The SCM_FLAG_NOTVALID flag is redundant with SCM_FLAG_VALID but exists
for compatibility with database values that were written by an older
version of RPSTIR.  A future commit will clear the SCM_FLAG_VALID flag
on upgrade if the SCM_FLAG_NOTVALID flag is set, which will make it
possible to eliminate the SCM_FLAG_NOTVALID flag entirely.

This change was made for the following reasons:

  * preserving the previous flag semantics would make it difficult to
    fix the evil twin bug

  * an evil twin certificate makes it possible to construct an invalid
    chain, so there is not much point in distinguishing between an
    incomplete chain and a complete but invalid chain

(An evil twin certificate is a maliciously crafted certificate with
the same public key, subject, and SKI as a victim's certificate.  The
point of an evil twin certificate is to make some of the victim's
objects look invalid by taking advantage of bugs in RP certificate
processing.)