eliminate verify_cert()'s chainOK parameter
Before, if there was no path to a trust anchor, verify_cert() set
*chainOK to false but returned success.  Now, verify_cert() simply
returns ERR_SCM_NOTVALID and there is no chainOK parameter.

An evil twin can make it appear as if there is a path to a trust
anchor, so there's no substantial difference between a certificate
without a path to a trust anchor and an invalid certificate with a
path to a trust anchor.